requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.2.1
colorama==0.4.6
comm==0.2.2
Cython==3.0.12
decorator==5.1.1
exceptiongroup==1.2.2
fastapi==0.115.14
filelock==3.18.0
fsspec==2025.3.0
h11==0.16.0
h5py==3.13.0
idna==3.4
imageio==2.37.0
ipykernel==6.29.5
jedi==0.19.1
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.1
pandas==2.2.3
Pillow==9.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pywin32==308
PyYAML==6.0.2
requests==2.28.1
scipy==1.15.3
six==1.16.0
sniffio==1.3.1
starlette==0.46.2
thop==0.1.1.post2209072238
torch==2.2.0+cu118
torch-vision==0.1.6.dev0
torchaudio==0.12.1+cu113
torchvision==0.17.0+cu118
tornado==6.4.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.154
urllib3==1.26.12
uvicorn==0.35.0
wcwidth==0.2.13
